refactor: extract server factory to src/server.ts for testability
